requirements.txt
# Streamlit and web framework
streamlit>=1.28.0
requests>=2.31.0

# Backend dependencies
python-multipart==0.0.6
python-dotenv==1.0.0
python-docx==1.0.1
PyPDF2==3.0.1
pydantic==2.4.2
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
openai==1.12.0
httpx==0.24.1
tiktoken==0.5.1
marshmallow==3.19.0
environs==9.5.0
chromadb==0.4.22
pymilvus==2.3.7
